refactor: move picker seletor hooks
