+++ src/PickerInput/RangePicker.tsx
@@ -33,7 +33,7 @@ import useRangePickerValue from './hooks/useRangePickerValue';
 import useRangeValue, { useInnerValue } from './hooks/useRangeValue';
 import useShowNow from './hooks/useShowNow';
 import Popup from './Popup';
-import RangeSelector, { type SelectorIdType } from '../NewPicker/PickerInput/Selector/RangeSelector';
+import RangeSelector, { type SelectorIdType } from './Selector/RangeSelector';
 
 function separateConfig<T>(config: T | [T, T] | null | undefined, defaultConfig: T): [T, T] {
   const singleConfig = config ?? defaultConfig;
