+++ src/PickerInput/Selector/Input.tsx
@@ -3,9 +3,9 @@ import { useEvent } from 'rc-util';
 import useLayoutEffect from 'rc-util/lib/hooks/useLayoutEffect';
 import raf from 'rc-util/lib/raf';
 import * as React from 'react';
-import { leftPad } from '../../../utils/miscUtil';
-import PickerContext from '../../../PickerInput/context';
-import useLockEffect from '../../../PickerInput/hooks/useLockEffect';
+import { leftPad } from '../../utils/miscUtil';
+import PickerContext from '../context';
+import useLockEffect from '../hooks/useLockEffect';
 import Icon from './Icon';
 import MaskFormat from './MaskFormat';
 import { getMaskRange } from './util';
